feat: add dockerfile

+++ src/auth/routes.ts
@@ -2,51 +2,70 @@ import { FastifyInstance, FastifyRequest } from "fastify";
 import { signup, signin } from "./interface.js";
 import { AuthUseCase } from "./usecase.js";
 import { app, fastifyPassport } from "../server.js";
+import { signinValidation } from "../validation/signin/schema.js";
+import { signupValidation } from "../validation/signup/schema.js";
+import { validatorCompiler } from "../validation/validator.js";
 
 export async function authRoutes(fastify: FastifyInstance) {
   const authUseCase = new AuthUseCase()
 
-  fastify.post<{ Body: signin }>('/signin', async (req, reply) => {
-    const { email, password } = req.body;
-
+  fastify.post<{ Body: signin }>('/signin', {
+    schema: signinValidation.schema,
+    validatorCompiler: validatorCompiler
+  }, async (req, reply) => {
     try {
+      const { email, password } = req.body;
       await authUseCase.signin({ email, password }, req);
       reply.status(200);
-
     } catch (error) {
-      reply.status(400).send(error);
+      reply.status(500).send(error);
     }
 
   });
 
-  fastify.post<{ Body: signup }>('/signup', async (req, reply) => {
-    const { email, password, name } = req.body
-
+  fastify.post<{ Body: signup }>('/signup', {
+    schema: signupValidation.schema,
+    validatorCompiler: validatorCompiler
+  }, async (req, reply) => {
     try {
+      const { email, password, name } = req.body
       await authUseCase.signup({ email, password, name });
-
     } catch (error) {
-      reply.status(400).send(error);
+      reply.status(400).send(error)
     }
 
   });
 
-  fastify.get('/microsoft', fastifyPassport.authenticate('azuread-openidconnect', { failureRedirect: '/api/v1/auth/auth-failure' }));
+  fastify.get('/microsoft', { preHandler: [app.setRedirectUrl] },
+    fastifyPassport.authenticate('azuread-openidconnect', { failureRedirect: '/api/v1/auth/auth-failure' }
+    ));
 
-  fastify.post('/callback',
+  fastify.get('/callback',
     { preValidation: fastifyPassport.authenticate('azuread-openidconnect', { failureRedirect: '/api/v1/auth/auth-failure' }) },
     (req, reply) => {
-      reply.redirect(`${process.env.FRONTEND_URL}`)
+
+      const { redirect_url_frontend } = req.cookies
+
+      if (!redirect_url_frontend) {
+        reply.code(400).send({ error: 'O cookie redirect_url_frontend está ausente' })
+        return;
+      }
+
+      reply.redirect(redirect_url_frontend)
     }
   )
 
-  fastify.get('/logout', async (req: FastifyRequest, reply) => {
+  fastify.get('/logout', { preHandler: [app.ensureAuthenticated] }, async (req: FastifyRequest, reply) => {
     try {
-      await req.logOut();
-      req.session.delete();
-      if (req.session.deleted) {
-        reply.redirect(`${process.env.FRONTEND_URL}`);
-        return
+      const { redirect_url_frontend } = req.cookies
+
+      if (redirect_url_frontend) {
+        await req.logOut();
+        req.session.delete();
+        if (req.session.deleted) {
+          reply.redirect(redirect_url_frontend)
+          return
+        }
       }
 
       reply.status(500).send('Erro ao tentar sair!')
@@ -60,7 +79,7 @@ export async function authRoutes(fastify: FastifyInstance) {
   });
 
   fastify.get('/auth-failure', async (req, reply) => {
-    reply.status(401).send({ message: 'Authentication failed' });
+    reply.code(500).send('Erro ao efetuar o login');
   })
-  
+
 }
